feat: 添加use_theme_config hook

+++ src/pages/select_history.rs
@@ -199,12 +199,12 @@ impl Component for SelectHistory {
                     Ok(None)
                 }
                 KeyCode::Enter => {
-                    if let Some(index) = self.state.selected {
-                        if self.confirm_state.is_confirm() {
-                            self.history.lock().await.remove_index(index);
-                            self.history.lock().await.save()?;
-                            self.state.select(None);
-                        }
+                    if let Some(index) = self.state.selected
+                        && self.confirm_state.is_confirm()
+                    {
+                        self.history.lock().await.remove_index(index);
+                        self.history.lock().await.save()?;
+                        self.state.select(None);
                     }
                     self.confirm_state.hide();
                     Ok(None)
